backend-project/user/models.py
import re
from django.contrib.auth.base_user import BaseUserManager
from django.db import models
from django.core import validators
from django.contrib.auth.models import AbstractBaseUser, PermissionsMixin
from reversion import revisions as reversion
